feat: fetching with snopshot

+++ src/index.js
@@ -1,5 +1,5 @@
 import { initializeApp } from "firebase/app";
-import { getFirestore, collection, getDocs, addDoc, deleteDoc, doc } from "firebase/firestore";
+import { getFirestore, collection, getDocs, addDoc, deleteDoc, doc, onSnapshot } from "firebase/firestore";
 
 const firebaseConfig = {
     apiKey: "AIzaSyBSZB-CmomA5mdjICWXAc9YQBH_OlSqQpc",
@@ -17,7 +17,10 @@ const db = getFirestore(app);
 
 const collectionRef = collection(db, "blogs");
 
-getDocs(collectionRef).then((data) => {
+// getDocs(collectionRef).then((data) => {
+
+// })
+onSnapshot(collectionRef, (data) => {
     let arr = [];
     data.forEach((doc) => {
         arr.push({ ...doc.data(), id: doc.id });
@@ -28,11 +31,19 @@ getDocs(collectionRef).then((data) => {
 const addForm = document.querySelector(".add");
 addForm.addEventListener('submit', (e) => {
     e.preventDefault();
+    if (!addForm.title.value || !addForm.desc.value) {
+        console.log("Empt form can't be submitted!")
+        return
+    }
     addDocs();
 })
 const deleteForm = document.querySelector(".delete");
 deleteForm.addEventListener('submit', (e) => {
     e.preventDefault();
+    if (!deleteForm.delete.value) {
+        console.log("Empt form can't be submitted!")
+        return
+    }
     deleteDoc(doc(db, 'blogs', deleteForm.delete.value)).then(data => {
         console.log("Data deleted", data)
     }).catch(err => {
